refactor(server): move recognizers_ja.py to server, drop training dep
server only imported `recognizers_ja.py` from `pleno-ner-training`. Keeping
that workspace dep meant any new training-side dependency (torch / nvidia
CUDA libs / openai) could leak into the server image and re-trigger the
fly.io 8GB rootfs unpack failure that #47 fought to fix.

Move the file to `server/src/recognizers_ja.py` (presidio-only, 368 lines)
and break the workspace edge:

- `server/src/app.py`: relative import `from .recognizers_ja import ...`
- `server/tests/conftest.py`: sys.path-injected bare import (server is a
  workspace member, not a true Python package — `from server.src.X` does
  not resolve under `cd server && pytest`)
- `server/pyproject.toml`: drop `pleno-ner-training` workspace dep
- `Dockerfile`: stop COPYing `packages/training/` into the image

Training side stays as a downstream consumer:

- `packages/training/tests/conftest.py` adds `server/src` to sys.path so
  the regex-correctness suite (`test_recognizers_ja.py`) keeps working
  without re-introducing the package edge
- `baselines_ja._build_presidio_predictor` uses the same lazy sys.path
  insert (only hit on the `[bench]` Presidio baseline path)
- `compare_baselines.RECOGNIZERS_PATH_REL` retargeted to the new location
  for git-blob fingerprinting

Tests:
- `cd server && uv run pytest tests/`: 91 passed, 101 skipped
- `uv run --directory packages/training pytest tests/test_recognizers_ja.py
   -k 'not benchmark and not f1'`: 25 passed, 4 deselected

Closes #74
