+++ server/pyproject.toml
@@ -10,11 +10,7 @@ dependencies = [
     "spacy[ja]>=3.8",
     "uvicorn[standard]>=0.35.0",
     "scalar-fastapi>=1.6.0",
-    "pleno-ner-training",
 ]
-
-[tool.uv.sources]
-pleno-ner-training = { workspace = true }
 
 [project.optional-dependencies]
 image = [
